requirements.txt
google-genai>=1.16.1
